Move form status functions to own file
