+++ packages/composite-checkout/src/components/checkout-provider.js
@@ -1,7 +1,7 @@
 /**
  * External dependencies
  */
-import React, { useReducer, useCallback, useContext, useEffect, useState, useRef } from 'react';
+import React, { useContext, useState, useRef } from 'react';
 import PropTypes from 'prop-types';
 import { ThemeProvider } from 'emotion-theming';
 import debugFactory from 'debug';
@@ -14,6 +14,7 @@ import CheckoutErrorBoundary from './checkout-error-boundary';
 import { LocalizeProvider } from '../lib/localize';
 import { LineItemsProvider } from '../lib/line-items';
 import { RegistryProvider, createRegistry } from '../lib/registry';
+import { useFormStatusManager } from '../lib/form-status';
 import defaultTheme from '../theme';
 import {
 	validateArg,
@@ -181,52 +182,3 @@ export const useCheckoutRedirects = () => {
 	}
 	return { successRedirectUrl, failureRedirectUrl };
 };
-
-export function useFormStatus() {
-	const { formStatus, setFormStatus } = useContext( CheckoutContext );
-	return [ formStatus, setFormStatus ];
-}
-
-function useFormStatusManager( isLoading ) {
-	const [ formStatus, dispatchFormStatus ] = useReducer(
-		formStatusReducer,
-		isLoading ? 'loading' : 'ready'
-	);
-	const setFormStatus = useCallback( payload => {
-		if ( typeof payload === 'function' ) {
-			return dispatchFormStatus( { type: 'FORM_STATUS_CHANGE_WITH_FUNCTION', payload } );
-		}
-		return dispatchFormStatus( { type: 'FORM_STATUS_CHANGE', payload } );
-	}, [] );
-	useEffect( () => {
-		const newStatus = isLoading ? 'loading' : 'ready';
-		debug( `isLoading has changed to ${ isLoading }; setting form status to ${ newStatus }` );
-		setFormStatus( newStatus );
-	}, [ isLoading, setFormStatus ] );
-	debug( `form status is ${ formStatus }` );
-	return [ formStatus, setFormStatus ];
-}
-
-function formStatusReducer( state, action ) {
-	switch ( action.type ) {
-		case 'FORM_STATUS_CHANGE':
-			validateStatus( action.payload );
-			debug( 'setting form status to', action.payload );
-			return action.payload;
-		case 'FORM_STATUS_CHANGE_WITH_FUNCTION': {
-			const newStatus = action.payload( state );
-			validateStatus( newStatus );
-			debug( 'setting form status to', newStatus );
-			return newStatus;
-		}
-		default:
-			return state;
-	}
-}
-
-function validateStatus( status ) {
-	const validStatuses = [ 'loading', 'ready', 'submitting' ];
-	if ( ! validStatuses.includes( status ) ) {
-		throw new Error( `Invalid form status '${ status }'` );
-	}
-}
